Fixed bug.
It now works all well with mins > 0 and maxs < 0, but I should test a little more, recomment,
and make things look nice. This was a little sloppy, but effective. I also need to readd forced
ticks at the endpoints for these strange bounds because, in a hurry to ship, I excluded them.